# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "CRM",
        "contact": {},
        "version": "1.0"
    },
    "host": "192.168.0.120:8000",
    "paths": {
...
        "/api/v1/hoa-don-xuat-kho": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Filter export invoice based on provided filters",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "export invoice"
                ],
                "summary": "Filter Export Invoice",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filters in JSON format",
                        "name": "filters",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order (asc/desc)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit per page",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update export invoice",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "export invoice"
                ],
                "summary": "Update Export Invoice",
                "parameters": [
                    {
                        "description": "Update Export Invoice Data",
                        "name": "Update_Export_Invoice",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.Hoa_don_xuat_kho_update"
                        }
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new export invoice entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "export invoice"
                ],
                "summary": "Create Export Invoice",
                "parameters": [
                    {
                        "description": "Export Invoice data",
                        "name": "Export_Invoice",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.Hoa_don_xuat_kho_create"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/hoa-don-xuat-kho/lock": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lock export invoice",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "export invoice"
                ],
                "summary": "Lock Export Invoice",
                "parameters": [
                    {
                        "description": "Lock Export Invoice Data",
                        "name": "Lock_Export_Invoice",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.Hoa_don_xuat_kho_lock"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/hoa-don-xuat-kho/tra-hang": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return Product for export invoice",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "export invoice"
                ],
                "summary": "Export Return Product",
                "parameters": [
                    {
                        "description": "Return Product Data",
                        "name": "Return_Product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.Tra_hang_xuat_kho_request"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/hoa-don-xuat-kho/tra-no": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Pay Debt for export invoice",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "export invoice"
                ],
                "summary": "Export Debt Payment",
                "parameters": [
                    {
                        "description": "Debt Payment Data",
                        "name": "Debt_Payment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.Tra_no_xuat_kho_request"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/khach-hang": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Filter customer based on provided filters",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "customer"
                ],
                "summary": "Filter Customer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filters in JSON format",
                        "name": "filters",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order (asc/desc)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit per page",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing customer entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customer"
                ],
                "summary": "Update Customer",
                "parameters": [
                    {
                        "description": "Updated customer data",
                        "name": "Customer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.Khach_hang_update"
                        }
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new customer entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customer"
                ],
                "summary": "Create Customer",
                "parameters": [
                    {
                        "description": "Customer data",
                        "name": "Discount_Type",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.Khach_hang_create"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/khach-hang/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete an existing customer entry",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customer"
                ],
                "summary": "Delete Customer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer ID to be deleted",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "requests.Chi_tiet_hoa_don_xuat_kho_create": {
            "type": "object",
            "required": [
                "chiet_khau",
                "ctsp_id",
                "don_vi_tinh",
                "ds_sku",
                "gia_ban",
                "gia_nhap",
                "la_qua_tang",
                "loi_nhuan",
                "san_pham_id",
                "sku",
                "so_luong_ban",
                "thanh_tien"
            ],
            "properties": {
                "chiet_khau": {
                    "type": "number"
                },
                "ctsp_id": {
                    "type": "integer"
                },
                "don_vi_tinh": {
                    "type": "string"
                },
                "ds_sku": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/requests.Chi_tiet_sku"
                    }
                },
                "gia_ban": {
                    "type": "number"
                },
                "gia_nhap": {
                    "type": "number"
                },
                "la_qua_tang": {
                    "type": "boolean"
                },
                "loi_nhuan": {
                    "type": "number"
                },
                "san_pham_id": {
                    "type": "integer"
                },
                "sku": {
                    "type": "string"
                },
                "so_luong_ban": {
                    "type": "integer"
                },
                "thanh_tien": {
                    "type": "number"
                }
            }
        },
...
        "requests.Chi_tiet_san_pham_xuat_tra": {
            "type": "object",
            "required": [
                "cthd_xuat_kho_id",
                "sku",
                "so_luong_tra"
            ],
            "properties": {
                "cthd_xuat_kho_id": {
                    "type": "integer"
                },
                "sku": {
                    "type": "string"
                },
                "so_luong_tra": {
                    "type": "integer"
                }
            }
        },
        "requests.Chi_tiet_sku": {
            "type": "object",
            "properties": {
                "gia_ban": {
                    "description": "Gia_nhap\t\t\tfloat32\t\t\t\t`json:\"gia_nhap\" binding:\"omitempty\"`",
                    "type": "number"
                },
                "sku": {
                    "type": "string"
                },
                "so_luong_ban": {
                    "type": "integer"
                }
            }
        },
...
        "requests.Hoa_don_xuat_kho_create": {
            "type": "object",
            "required": [
                "da_giao_hang",
                "ds_san_pham_xuat",
                "gia_tri_chiet_khau",
                "khach_hang_id",
                "loai_chiet_khau",
                "loi_nhuan",
                "ngay_xuat",
                "nhan_vien_giao_hang_id",
                "nhan_vien_sale_id",
                "thanh_tien",
                "tong_tien",
                "tra_truoc",
                "vat"
            ],
            "properties": {
                "con_lai": {
                    "type": "number"
                },
                "da_giao_hang": {
                    "type": "boolean"
                },
                "ds_san_pham_xuat": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/requests.Chi_tiet_hoa_don_xuat_kho_create"
                    }
                },
                "ghi_chu": {
                    "type": "string"
                },
                "gia_tri_chiet_khau": {
                    "type": "integer"
                },
                "khach_hang_id": {
                    "type": "integer"
                },
                "loai_chiet_khau": {
                    "type": "integer"
                },
                "loi_nhuan": {
                    "description": "Tong_gia_nhap\t\t\t\tfloat32\t\t\t\t\t\t`json:\"tong_gia_nhap\" binding:\"omitempty\"`",
                    "type": "number"
                },
                "ma_hoa_don": {
                    "type": "string"
                },
                "ngay_xuat": {
                    "type": "string"
                },
                "nhan_vien_giao_hang_id": {
                    "type": "integer"
                },
                "nhan_vien_sale_id": {
                    "type": "integer"
                },
                "so_hoa_don": {
                    "type": "integer"
                },
                "thanh_tien": {
                    "type": "number"
                },
                "tong_tien": {
                    "type": "number"
                },
                "tra_truoc": {
                    "type": "number"
                },
                "vat": {
                    "type": "number"
                }
            }
        },
        "requests.Hoa_don_xuat_kho_lock": {
            "type": "object",
            "required": [
                "hoa_don_id"
            ],
            "properties": {
                "hoa_don_id": {
                    "type": "integer"
                },
                "lock_or_open": {
                    "type": "string",
                    "enum": [
                        "lock",
                        "open"
                    ]
                }
            }
        },
        "requests.Hoa_don_xuat_kho_update": {
            "type": "object",
            "required": [
                "hoa_don_id",
                "khach_hang_id",
                "ngay_xuat",
                "nhan_vien_giao_hang_id",
                "nhan_vien_sale_id"
            ],
            "properties": {
                "da_giao_hang": {
                    "type": "boolean"
                },
                "ghi_chu": {
                    "type": "string"
                },
                "gia_tri_chiet_khau": {
                    "type": "integer"
                },
                "hoa_don_id": {
                    "type": "integer"
                },
                "khach_hang_id": {
                    "type": "integer"
                },
                "loai_chiet_khau": {
                    "type": "integer"
                },
                "ngay_xuat": {
                    "type": "string"
                },
                "nhan_vien_giao_hang_id": {
                    "type": "integer"
                },
                "nhan_vien_sale_id": {
                    "type": "integer"
                },
                "tra_truoc": {
                    "type": "number"
                },
                "vat": {
                    "type": "number"
                }
            }
        },
        "requests.Khach_hang_create": {
            "type": "object",
            "required": [
                "dia_chi",
                "dien_thoai",
                "ho_ten"
            ],
            "properties": {
                "dia_chi": {
                    "type": "string"
                },
                "dien_thoai": {
                    "type": "string"
                },
                "ho_ten": {
                    "type": "string"
                }
            }
        },
        "requests.Khach_hang_update": {
            "type": "object",
            "required": [
                "dia_chi",
                "dien_thoai",
                "ho_ten",
                "id"
            ],
            "properties": {
                "dia_chi": {
                    "type": "string"
                },
                "dien_thoai": {
                    "type": "string"
                },
                "ho_ten": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
        "requests.Tra_hang_xuat_kho_request": {
            "type": "object",
            "required": [
                "ds_san_pham_tra",
                "hoa_don_id"
            ],
            "properties": {
                "ds_san_pham_tra": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/requests.Chi_tiet_san_pham_xuat_tra"
                    }
                },
                "hoa_don_id": {
                    "type": "integer"
                }
            }
        },
...
        "requests.Tra_no_xuat_kho_request": {
            "type": "object",
            "required": [
                "hoa_don_id",
                "tien_tra"
            ],
            "properties": {
                "hoa_don_id": {
                    "type": "integer"
                },
                "tien_tra": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}